Review Field Made decimal
Made the review field decimal from integer. Because if there are two reviews the average of them will not be always as an integer value.
like (3+4)/2=3.5 review points

+++ BookLender/mainapp/models.py
@@ -13,7 +13,7 @@ class UserProfile(models.Model):
     current_location = models.CharField('Current Location', max_length=255, null=False, default='default')
     phone_number = models.CharField('Phone Number', max_length=255, null=False, default='default')
     birth_date = models.DateField('Birth Date', null=False, default=date(2000, 1, 1))
-    review = models.IntegerField('Review Score', null=True)
+    review = models.DecimalField('Review Score', max_digits=3, decimal_places=1, null=True)
 
     def __str__(self):
         return self.user.username
